Extract Feature to its own file
